Moving ClientGenerator/ServerGenerator out into their own packages
