+++ modules/java-dropwizard/src/main/scala/dev/guardrail/generators/Java/DropwizardServerGenerator.scala
@@ -22,7 +22,8 @@ import dev.guardrail.protocol.terms.server._
 import dev.guardrail.shims.OperationExt
 import dev.guardrail.terms.collections.CollectionsAbstraction
 import dev.guardrail.terms.{ CollectionsLibTerms, RouteMeta, SecurityScheme }
-import dev.guardrail.{ CustomExtractionField, RenderedRoutes, StrictProtocolElems, SupportDefinition, Target, TracingField }
+import dev.guardrail.{ StrictProtocolElems, SupportDefinition, Target }
+import dev.guardrail.generators.{ CustomExtractionField, RenderedRoutes, TracingField }
 import io.swagger.v3.oas.models.Operation
 import scala.compat.java8.OptionConverters._
 import scala.concurrent.Future
