+++ modules/scala-http4s/src/main/scala/dev/guardrail/generators/Scala/Http4sServerGenerator.scala
@@ -4,10 +4,10 @@ import cats.Monad
 import cats.data.NonEmptyList
 import cats.implicits._
 import cats.Traverse
-import dev.guardrail.{ CustomExtractionField, RenderedRoutes, StrictProtocolElems, Target, TracingField, UserError }
+import dev.guardrail.{ StrictProtocolElems, Target, UserError }
 import dev.guardrail.core.{PathExtractor, Tracker}
 import dev.guardrail.core.extract.{ ServerRawResponse, TracingLabel }
-import dev.guardrail.generators.{ LanguageParameter, LanguageParameters }
+import dev.guardrail.generators.{ CustomExtractionField, LanguageParameter, LanguageParameters, RenderedRoutes, TracingField }
 import dev.guardrail.generators.syntax._
 import dev.guardrail.generators.operations.TracingLabelFormatter
 import dev.guardrail.generators.syntax.Scala._
